use zod for edit and delete

+++ src/actions/actions.ts
@@ -3,11 +3,11 @@
 import prisma from '@/lib/db';
 import { PetEssentials } from '@/lib/types';
 import { sleep } from '@/lib/utils';
-import { petFormSchema } from '@/lib/validations';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 import { Pet } from '@prisma/client';
 import { revalidatePath } from 'next/cache';
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
 
   const validatedPet = petFormSchema.safeParse(pet);
@@ -30,15 +30,24 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], newPet: PetEssentials) {
+export async function editPet(petId: unknown, newPet: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPet);
+
+  if (!validatedPet.success || !validatedPetId.success) {
+    return {
+      message: "Invalid pet data."
+    }
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -49,12 +58,21 @@ export async function editPet(petId: Pet['id'], newPet: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data."
+    }
+  }
+
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
